+++ RELEASENOTES.md
@@ -55,6 +55,7 @@ The following binary downloads have been compiled with ESP8266/Arduino library c
 - Change Settings text handling allowing variable length text within a total text pool of 699 characters
 - Change Smoother ``Fade`` using 100Hz instead of 20Hz animation (#7179)
 - Change number of rule ``Var``s and ``Mem``s from 5 to 16 (#4933)
+- Change number of ``FriendlyName``s from 4 to 8
 - Add support for max 150 characters in most command parameter strings (#3686, #4754)
 - Add support for GPS as NTP server by Christian Baars and Adrian Scillato
 - Add support for ``AdcParam`` parameters to control ADC0 Moisture formula by Federico Leoni (#7309)